pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "EXACT"                  # Your package name on PyPI
version = "0.1.0"                  # Semantic version
description = "Framework-agnostic library for explainable AI (XAI) methods like GradCAM, LIME, SHAP."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Glenn Mathews", email = "glennmathews18@gmail.com" }
]
keywords = ["xai", "explainability", "gradcam", "lime", "shap", "ml", "ai"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
requires-python = ">=3.8"

# Runtime dependencies (minimal)
dependencies = [
    "numpy",
    "matplotlib",
    "pillow",
]
                                                                  
[project.optional-dependencies]
pytorch = ["torch", "torchvision"]
tensorflow = ["tensorflow", "keras"]
all = ["torch", "torchvision", "tensorflow", "keras"]  

[project.urls]
#Homepage = "https://github.com/yourusername/mlmodels"
#Documentation = "https://github.com/yourusername/mlmodels/wiki"
#Source = "https://github.com/yourusername/mlmodels"
#Tracker = "https://github.com/yourusername/mlmodels/issues"

[tool.setuptools]
package-dir = {"" = "src"}   # tells setuptools where your code lives

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]         # don’t ship tests to PyPI